requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-nested-admin==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
Faker==33.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
